step2 - refactor
